First draft of a login process
I based myself on the "auth" example of express.js (mainly the .ejs
files). Went throught it so I could understand what was being done. From
there, I added connection to database.
Its possible to sign in and sign up, and there´s session support, so if
you close the tab and open the address in a new one, it will leave u
logged in.