[Auth] Session Storage & Cookie support
Added session storage and cookies support. Sessions are now being
storaged in a Redis Store (a NoSQL DB - theres a reason not to put in
the Postgres - even though its possible).

With cookies, the user can close the web-browser, and when they are
back, the session will still be valid (if the browser is not set-up to
clean cookies when closed, ofc)

Changed some code as well, aiming to make it more modular and re-usable.
I would say we already have a functional log-in module. From now on, we
just have to add as desired (I guess)