Final structure
This is the structure that we should be using from now on. Development,
will be done through the develop branch, and fully working commits
should be done to the master.

No folder should be deleted from now on, but news can be added if its
necessary for the module you are developing.

Furthermore, the node_modules folder is ignored, so please if you are
installing a new module, add it into the package.json and put a
disclaimer on the commit so, whoever is syncing does not forget to "npm
install" before developing again.