Authentication Module
This is a working authentication module which includes email dispatching
functionality. Before anything, please run "npm install" from the root
folder, so it installs the necessary additional modules. There are some
files which are ignored when uploaded. I will make them available on
Dropbox.

Im working directly with default http responses (200, 400, 500) or
redirects. However, I had to create some minimal views in order to test
it. What will be in the views are not so important, as long as the name
of them are the same (and the fields when applicable).

Any doubts, please just ask me (trello or fb)